+++ pkg/server/router/credential_test.go
@@ -7,11 +7,13 @@ import (
 	"time"
 
 	credsdk "github.com/TBD54566975/ssi-sdk/credential"
-	"github.com/goccy/go-json"
+	"github.com/TBD54566975/ssi-sdk/credential/signing"
+	"github.com/TBD54566975/ssi-sdk/crypto"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/tbd54566975/ssi-service/config"
 	"github.com/tbd54566975/ssi-service/pkg/service/credential"
+	"github.com/tbd54566975/ssi-service/pkg/service/did"
 	"github.com/tbd54566975/ssi-service/pkg/service/framework"
 	"github.com/tbd54566975/ssi-service/pkg/storage"
 )
@@ -43,6 +45,7 @@ func TestCredentialRouter(t *testing.T) {
 
 		serviceConfig := config.CredentialServiceConfig{BaseServiceConfig: &config.BaseServiceConfig{Name: "credential"}}
 		keyStoreService := testKeyStoreService(tt, bolt)
+		didService := testDIDService(tt, bolt, keyStoreService)
 		credService, err := credential.NewCredentialService(serviceConfig, bolt, keyStoreService)
 		assert.NoError(tt, err)
 		assert.NotEmpty(tt, credService)
@@ -52,7 +55,12 @@ func TestCredentialRouter(t *testing.T) {
 		assert.Equal(tt, framework.StatusReady, credService.Status().Status)
 
 		// create a credential
-		issuer := "did:test:123"
+
+		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{Method: did.KeyMethod, KeyType: crypto.Ed25519})
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, issuerDID)
+
+		issuer := issuerDID.DID.ID
 		subject := "did:test:345"
 		createdCred, err := credService.CreateCredential(credential.CreateCredentialRequest{
 			Issuer:  issuer,
@@ -65,25 +73,25 @@ func TestCredentialRouter(t *testing.T) {
 		})
 		assert.NoError(tt, err)
 		assert.NotEmpty(tt, createdCred)
-		assert.NotEmpty(tt, createdCred.Credential)
+		assert.NotEmpty(tt, createdCred.CredentialJWT)
+
+		parsedCredential, err := signing.ParseVerifiableCredentialFromJWT(*createdCred.CredentialJWT)
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedCredential)
 
 		// make sure it has the right data
-		assert.Equal(tt, issuer, createdCred.Credential.Issuer)
-		assert.Equal(tt, subject, createdCred.Credential.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
-		assert.Equal(tt, "Satoshi", createdCred.Credential.CredentialSubject["firstName"])
-		assert.Equal(tt, "Nakamoto", createdCred.Credential.CredentialSubject["lastName"])
+		assert.Equal(tt, issuer, parsedCredential.Issuer)
+		assert.Equal(tt, subject, parsedCredential.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
+		assert.Equal(tt, "Satoshi", parsedCredential.CredentialSubject["firstName"])
+		assert.Equal(tt, "Nakamoto", parsedCredential.CredentialSubject["lastName"])
 
 		// get it back
-		gotCred, err := credService.GetCredential(credential.GetCredentialRequest{ID: createdCred.Credential.ID})
+		gotCred, err := credService.GetCredential(credential.GetCredentialRequest{ID: parsedCredential.ID})
 		assert.NoError(tt, err)
 		assert.NotEmpty(tt, gotCred)
 
-		// marshal to JSON and compare for object equality
-		createdBytes, err := json.Marshal(createdCred.Credential)
-		assert.NoError(tt, err)
-		gotBytes, err := json.Marshal(gotCred.Credential)
-		assert.NoError(tt, err)
-		assert.Equal(tt, createdBytes, gotBytes)
+		// compare for object equality
+		assert.Equal(tt, createdCred.CredentialJWT, gotCred.CredentialJWT)
 
 		// get a cred that doesn't exist
 		_, err = credService.GetCredential(credential.GetCredentialRequest{ID: "bad"})
@@ -93,23 +101,41 @@ func TestCredentialRouter(t *testing.T) {
 		// get by schema - no schema
 		bySchema, err := credService.GetCredentialsBySchema(credential.GetCredentialBySchemaRequest{Schema: ""})
 		assert.NoError(tt, err)
-		assert.Len(tt, bySchema.Credentials, 1)
-		assert.Equal(tt, bySchema.Credentials[0].ID, createdCred.Credential.ID)
-		assert.EqualValues(tt, bySchema.Credentials[0].CredentialSchema, createdCred.Credential.CredentialSchema)
+		assert.Len(tt, bySchema.CredentialJWTs, 1)
+
+		// parse cred
+		parsedBySchema, err := signing.ParseVerifiableCredentialFromJWT(bySchema.CredentialJWTs[0])
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedBySchema)
+
+		assert.Equal(tt, parsedCredential.ID, parsedBySchema.ID)
+		assert.EqualValues(tt, parsedCredential.CredentialSchema, parsedBySchema.CredentialSchema)
 
 		// get by subject
 		bySubject, err := credService.GetCredentialsBySubject(credential.GetCredentialBySubjectRequest{Subject: subject})
 		assert.NoError(tt, err)
-		assert.Len(tt, bySubject.Credentials, 1)
-		assert.Equal(tt, bySubject.Credentials[0].ID, createdCred.Credential.ID)
-		assert.Equal(tt, bySubject.Credentials[0].CredentialSubject[credsdk.VerifiableCredentialIDProperty], createdCred.Credential.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
+		assert.Len(tt, bySubject.CredentialJWTs, 1)
+
+		// parse cred
+		parsedBySubject, err := signing.ParseVerifiableCredentialFromJWT(bySubject.CredentialJWTs[0])
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedBySubject)
+
+		assert.Equal(tt, parsedCredential.ID, parsedBySubject.ID)
+		assert.Equal(tt, parsedCredential.CredentialSubject[credsdk.VerifiableCredentialIDProperty], parsedBySubject.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
 
 		// get by issuer
 		byIssuer, err := credService.GetCredentialsByIssuer(credential.GetCredentialByIssuerRequest{Issuer: issuer})
 		assert.NoError(tt, err)
-		assert.Len(tt, byIssuer.Credentials, 1)
-		assert.Equal(tt, byIssuer.Credentials[0].ID, createdCred.Credential.ID)
-		assert.Equal(tt, byIssuer.Credentials[0].Issuer, createdCred.Credential.Issuer)
+		assert.Len(tt, byIssuer.CredentialJWTs, 1)
+
+		// parse cred
+		parsedByIssuer, err := signing.ParseVerifiableCredentialFromJWT(byIssuer.CredentialJWTs[0])
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedByIssuer)
+
+		assert.Equal(tt, parsedCredential.ID, parsedByIssuer.ID)
+		assert.Equal(tt, parsedCredential.Issuer, parsedByIssuer.Issuer)
 
 		// create another cred with the same issuer, different subject, different schema
 		anotherCreatedCred, err := credService.CreateCredential(credential.CreateCredentialRequest{
@@ -127,32 +153,42 @@ func TestCredentialRouter(t *testing.T) {
 		// get by issuer
 		byIssuer, err = credService.GetCredentialsByIssuer(credential.GetCredentialByIssuerRequest{Issuer: issuer})
 		assert.NoError(tt, err)
-		assert.Len(tt, byIssuer.Credentials, 2)
+		assert.Len(tt, byIssuer.CredentialJWTs, 2)
 
 		// make sure the schema and subject queries are consistent
 		bySchema, err = credService.GetCredentialsBySchema(credential.GetCredentialBySchemaRequest{Schema: ""})
 		assert.NoError(tt, err)
-		assert.Len(tt, bySchema.Credentials, 1)
-		assert.Equal(tt, bySchema.Credentials[0].ID, createdCred.Credential.ID)
-		assert.EqualValues(tt, bySchema.Credentials[0].CredentialSchema, createdCred.Credential.CredentialSchema)
+		assert.Len(tt, bySchema.CredentialJWTs, 1)
+
+		parsedBySchema, err = signing.ParseVerifiableCredentialFromJWT(bySchema.CredentialJWTs[0])
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedBySchema)
+
+		assert.Equal(tt, parsedCredential.ID, parsedBySchema.ID)
+		assert.EqualValues(tt, parsedCredential.CredentialSchema, parsedBySchema.CredentialSchema)
 
 		bySubject, err = credService.GetCredentialsBySubject(credential.GetCredentialBySubjectRequest{Subject: subject})
 		assert.NoError(tt, err)
-		assert.Len(tt, bySubject.Credentials, 1)
-		assert.Equal(tt, bySubject.Credentials[0].ID, createdCred.Credential.ID)
-		assert.Equal(tt, bySubject.Credentials[0].CredentialSubject[credsdk.VerifiableCredentialIDProperty], createdCred.Credential.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
+		assert.Len(tt, bySubject.CredentialJWTs, 1)
+
+		parsedBySubject, err = signing.ParseVerifiableCredentialFromJWT(bySubject.CredentialJWTs[0])
+		assert.NoError(tt, err)
+		assert.NotEmpty(tt, parsedBySubject)
+
+		assert.Equal(tt, parsedCredential.ID, parsedBySubject.ID)
+		assert.Equal(tt, parsedCredential.CredentialSubject[credsdk.VerifiableCredentialIDProperty], parsedBySubject.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
 
 		// delete a cred that doesn't exist (no error since idempotent)
 		err = credService.DeleteCredential(credential.DeleteCredentialRequest{ID: "bad"})
 		assert.NoError(tt, err)
 
 		// delete a credential that does exist
-		err = credService.DeleteCredential(credential.DeleteCredentialRequest{ID: createdCred.Credential.ID})
+		err = credService.DeleteCredential(credential.DeleteCredentialRequest{ID: parsedCredential.ID})
 		assert.NoError(tt, err)
 
 		// get it back
-		_, err = credService.GetCredential(credential.GetCredentialRequest{ID: createdCred.Credential.ID})
+		_, err = credService.GetCredential(credential.GetCredentialRequest{ID: parsedCredential.ID})
 		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), fmt.Sprintf("credential not found with id: %s", createdCred.Credential.ID))
+		assert.Contains(tt, err.Error(), fmt.Sprintf("credential not found with id: %s", parsedCredential.ID))
 	})
 }
