+++ pkg/server/server_test.go
@@ -3,7 +3,6 @@ package server
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -12,17 +11,14 @@ import (
 	"time"
 
 	"github.com/TBD54566975/ssi-sdk/credential/exchange"
-	"github.com/TBD54566975/ssi-sdk/credential/signing"
 
 	"github.com/tbd54566975/ssi-service/pkg/service/dwn"
 
-	credsdk "github.com/TBD54566975/ssi-sdk/credential"
 	manifestsdk "github.com/TBD54566975/ssi-sdk/credential/manifest"
 	"github.com/TBD54566975/ssi-sdk/crypto"
 	"github.com/dimfeld/httptreemux/v5"
 	"github.com/goccy/go-json"
 	"github.com/google/uuid"
-	"github.com/mr-tron/base58"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -93,1409 +89,6 @@ func TestReadinessAPI(t *testing.T) {
 
 	assert.Equal(t, svcframework.StatusReady, resp.Status.Status)
 	assert.Len(t, resp.ServiceStatuses, 0)
-}
-
-func TestDIDAPI(t *testing.T) {
-	t.Run("Test Get DID Methods", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		_, keyStoreService := testKeyStore(tt, bolt)
-		didService := testDIDRouter(tt, bolt, keyStoreService)
-
-		// get DID methods
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/dids", nil)
-		w := httptest.NewRecorder()
-
-		err = didService.GetDIDMethods(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-		assert.Equal(tt, http.StatusOK, w.Result().StatusCode)
-
-		var resp router.GetDIDMethodsResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.Len(tt, resp.DIDMethods, 1)
-		assert.Equal(tt, resp.DIDMethods[0], did.KeyMethod)
-	})
-
-	t.Run("Test Create DID By Method: Key", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		_, keyStoreService := testKeyStore(tt, bolt)
-		didService := testDIDRouter(tt, bolt, keyStoreService)
-
-		// create DID by method - key - missing body
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", nil)
-		w := httptest.NewRecorder()
-		params := map[string]string{
-			"method": "key",
-		}
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "invalid create DID request")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// with body, bad key type
-		createDIDRequest := router.CreateDIDByMethodRequest{KeyType: "bad"}
-		requestReader := newRequestValue(tt, createDIDRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", requestReader)
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not create DID for method<key> with key type: bad")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// with body, good key type
-		createDIDRequest = router.CreateDIDByMethodRequest{KeyType: crypto.Ed25519}
-		requestReader = newRequestValue(tt, createDIDRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", requestReader)
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.Contains(tt, resp.DID.ID, did.KeyMethod)
-	})
-
-	t.Run("Test Get DID By Method", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		_, keyStore := testKeyStore(tt, bolt)
-		didService := testDIDRouter(tt, bolt, keyStore)
-
-		// get DID by method
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/dids/bad/worse", nil)
-		w := httptest.NewRecorder()
-
-		// bad params
-		badParams := map[string]string{
-			"method": "bad",
-			"id":     "worse",
-		}
-		err = didService.GetDIDByMethod(newRequestContextWithParams(badParams), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get DID for method<bad>")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// good method, bad id
-		badParams1 := map[string]string{
-			"method": "key",
-			"id":     "worse",
-		}
-		err = didService.GetDIDByMethod(newRequestContextWithParams(badParams1), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get DID for method<key> with id: worse")
-
-		// reset recorder between calls
-		w.Flush()
-		// store a DID
-		createDIDRequest := router.CreateDIDByMethodRequest{KeyType: crypto.Ed25519}
-		requestReader := newRequestValue(tt, createDIDRequest)
-		params := map[string]string{"method": "key"}
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", requestReader)
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.NoError(tt, err)
-
-		var createdDID router.CreateDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&createdDID)
-		assert.NoError(tt, err)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get it back
-		createdID := createdDID.DID.ID
-		getDIDPath := fmt.Sprintf("https://ssi-service.com/v1/dids/key/%s", createdID)
-		req = httptest.NewRequest(http.MethodGet, getDIDPath, nil)
-
-		// good params
-		goodParams := map[string]string{
-			"method": "key",
-			"id":     createdID,
-		}
-		err = didService.GetDIDByMethod(newRequestContextWithParams(goodParams), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.GetDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.Equal(tt, createdID, resp.DID.ID)
-	})
-
-	t.Run("Test Get DIDs By Method", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		_, keyStore := testKeyStore(tt, bolt)
-		didService := testDIDRouter(tt, bolt, keyStore)
-
-		// get DIDs by method
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/dids/bad", nil)
-		w := httptest.NewRecorder()
-
-		// bad params
-		badParams := map[string]string{
-			"method": "bad",
-		}
-		err = didService.GetDIDsByMethod(newRequestContextWithParams(badParams), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get DIDs for method: bad")
-
-		// good method
-		goodParams := map[string]string{
-			"method": "key",
-		}
-		err = didService.GetDIDsByMethod(newRequestContextWithParams(goodParams), w, req)
-		assert.NoError(tt, err)
-		var gotDIDs router.GetDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&gotDIDs)
-		assert.NoError(tt, err)
-		assert.Empty(tt, gotDIDs)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// store two DIDs
-		createDIDRequest := router.CreateDIDByMethodRequest{KeyType: crypto.Ed25519}
-		requestReader := newRequestValue(tt, createDIDRequest)
-		params := map[string]string{"method": "key"}
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", requestReader)
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.NoError(tt, err)
-
-		var createdDID router.CreateDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&createdDID)
-		assert.NoError(tt, err)
-
-		// reset recorder between calls
-		w.Flush()
-
-		requestReader = newRequestValue(tt, createDIDRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dids/key", requestReader)
-
-		err = didService.CreateDIDByMethod(newRequestContextWithParams(params), w, req)
-		assert.NoError(tt, err)
-
-		var createdDID2 router.CreateDIDByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&createdDID2)
-		assert.NoError(tt, err)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get all dids for method
-
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/dids/key", requestReader)
-		err = didService.GetDIDsByMethod(newRequestContextWithParams(params), w, req)
-		assert.NoError(tt, err)
-
-		var gotDIDsResponse router.GetDIDsByMethodResponse
-		err = json.NewDecoder(w.Body).Decode(&gotDIDsResponse)
-		assert.NoError(tt, err)
-
-		knownDIDs := map[string]bool{createdDID.DID.ID: true, createdDID2.DID.ID: true}
-		for _, did := range gotDIDsResponse.DIDs {
-			if _, ok := knownDIDs[did.ID]; !ok {
-				tt.Error("got unknown DID")
-			} else {
-				delete(knownDIDs, did.ID)
-			}
-		}
-		assert.Len(tt, knownDIDs, 0)
-	})
-}
-
-func TestSchemaAPI(t *testing.T) {
-	t.Run("Test Create Schema", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		schemaService := testSchemaRouter(tt, bolt, keyStoreService)
-
-		simpleSchema := map[string]interface{}{
-			"type": "object",
-			"properties": map[string]interface{}{
-				"foo": map[string]interface{}{
-					"type": "string",
-				},
-			},
-			"required":             []interface{}{"foo"},
-			"additionalProperties": false,
-		}
-		badSchemaRequest := router.CreateSchemaRequest{Schema: simpleSchema}
-		schemaRequestValue := newRequestValue(tt, badSchemaRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/schemas", schemaRequestValue)
-		w := httptest.NewRecorder()
-
-		err = schemaService.CreateSchema(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "invalid create schema request")
-
-		// reset the http recorder
-		w.Flush()
-
-		schemaRequest := router.CreateSchemaRequest{Author: "did:test", Name: "test schema", Schema: simpleSchema}
-		schemaRequestValue = newRequestValue(tt, schemaRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/schemas", schemaRequestValue)
-		err = schemaService.CreateSchema(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateSchemaResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, resp.ID)
-		assert.EqualValues(tt, schemaRequest.Schema, resp.Schema.Schema)
-	})
-
-	t.Run("Test Get Schemas", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		schemaService := testSchemaRouter(tt, bolt, keyStoreService)
-
-		// get schema that doesn't exist
-		w := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/schemas/bad", nil)
-		err = schemaService.GetSchema(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "cannot get schema without ID parameter")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get schema with invalid id
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/schemas/bad", nil)
-		err = schemaService.GetSchema(newRequestContextWithParams(map[string]string{"id": "bad"}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get schema with id: bad")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get all schemas - get none
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/schemas", nil)
-		err = schemaService.GetSchemas(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-		var getSchemasResp router.GetSchemasResponse
-		err = json.NewDecoder(w.Body).Decode(&getSchemasResp)
-		assert.NoError(tt, err)
-		assert.Len(tt, getSchemasResp.Schemas, 0)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// create a schema
-		simpleSchema := map[string]interface{}{
-			"type": "object",
-			"properties": map[string]interface{}{
-				"foo": map[string]interface{}{
-					"type": "string",
-				},
-			},
-			"required":             []interface{}{"foo"},
-			"additionalProperties": false,
-		}
-		schemaRequest := router.CreateSchemaRequest{Author: "did:test", Name: "test schema", Schema: simpleSchema}
-		schemaRequestValue := newRequestValue(tt, schemaRequest)
-		createReq := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/schemas", schemaRequestValue)
-
-		err = schemaService.CreateSchema(newRequestContext(), w, createReq)
-		assert.NoError(tt, err)
-
-		var createResp router.CreateSchemaResponse
-		err = json.NewDecoder(w.Body).Decode(&createResp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, createResp.ID)
-		assert.EqualValues(tt, schemaRequest.Schema, createResp.Schema.Schema)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get it back
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/schemas/%s", createResp.ID), nil)
-		err = schemaService.GetSchema(newRequestContextWithParams(map[string]string{"id": createResp.ID}), w, req)
-		assert.NoError(tt, err)
-
-		var gotSchemaResp router.GetSchemaResponse
-		err = json.NewDecoder(w.Body).Decode(&gotSchemaResp)
-		assert.NoError(tt, err)
-
-		assert.Equal(tt, createResp.ID, gotSchemaResp.Schema.ID)
-		assert.Equal(tt, createResp.Schema.Schema, gotSchemaResp.Schema.Schema)
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get all schemas - get none
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/schemas", nil)
-		err = schemaService.GetSchemas(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-		err = json.NewDecoder(w.Body).Decode(&getSchemasResp)
-		assert.NoError(tt, err)
-		assert.Len(tt, getSchemasResp.Schemas, 1)
-	})
-}
-
-func TestCredentialAPI(t *testing.T) {
-	t.Run("Test Create Credential", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		// missing required field: data
-		badCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Expiry:  time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		badRequestValue := newRequestValue(tt, badCredRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", badRequestValue)
-		w := httptest.NewRecorder()
-
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "invalid create credential request")
-
-		// reset the http recorder
-		w.Flush()
-
-		// missing known issuer request
-		missingIssuerRequest := router.CreateCredentialRequest{
-			Issuer:  "did:abc:123",
-			Subject: "did:abc:456",
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		missingIssuerRequestValue := newRequestValue(tt, missingIssuerRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", missingIssuerRequestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get key for signing credential with key<did:abc:123>")
-
-		// reset the http recorder
-		w.Flush()
-
-		// good request
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, resp.CredentialJWT)
-		parsedCred, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCred)
-		assert.Equal(tt, parsedCred.Issuer, issuerDID.DID.ID)
-	})
-
-	t.Run("Test Get Credential By ID", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		w := httptest.NewRecorder()
-
-		// get a cred that doesn't exit
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/credentials/bad", nil)
-		err = credService.GetCredential(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "cannot get credential without ID parameter")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get a cred with an invalid id parameter
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/credentials/bad", nil)
-		err = credService.GetCredential(newRequestContextWithParams(map[string]string{"id": "bad"}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get credential with id: bad")
-
-		// reset recorder between calls
-		w.Flush()
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		// We expect a JWT credential
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.Empty(tt, resp.Credential)
-		assert.NotEmpty(tt, resp.CredentialJWT)
-
-		parsedCred, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCred)
-
-		// get credential by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credentials/%s", parsedCred.ID), nil)
-		err = credService.GetCredential(newRequestContextWithParams(map[string]string{"id": parsedCred.ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getCredResp router.GetCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&getCredResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getCredResp)
-		assert.NotEmpty(tt, getCredResp.CredentialJWT)
-
-		parsedCredResponse, err := signing.ParseVerifiableCredentialFromJWT(*getCredResp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredResponse)
-		assert.Equal(tt, parsedCredResponse.ID, getCredResp.ID)
-	})
-
-	t.Run("Test Get Credential By Schema", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		w := httptest.NewRecorder()
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		schemaID := "https://test-schema.com/name"
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Schema:  schemaID,
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, resp.CredentialJWT)
-
-		parsedCredResponse, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredResponse)
-
-		w.Flush()
-
-		// get credential by schema
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credential?schema=%s", schemaID), nil)
-		err = credService.GetCredentials(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var getCredsResp router.GetCredentialsResponse
-		err = json.NewDecoder(w.Body).Decode(&getCredsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getCredsResp)
-		assert.Len(tt, getCredsResp.CredentialJWTs, 1)
-
-		parsedCredsResponse, err := signing.ParseVerifiableCredentialFromJWT(getCredsResp.CredentialJWTs[0])
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredsResponse)
-
-		assert.Equal(tt, parsedCredResponse.ID, parsedCredsResponse.ID)
-		assert.Equal(tt, parsedCredResponse.CredentialSchema.ID, parsedCredsResponse.CredentialSchema.ID)
-	})
-
-	t.Run("Test Get Credential By Issuer", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		w := httptest.NewRecorder()
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		schemaID := "https://test-schema.com/name"
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Schema:  schemaID,
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, resp.CredentialJWT)
-
-		parsedCredResponse, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredResponse)
-
-		w.Flush()
-
-		// get credential by issuer id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credential?issuer=%s", issuerDID.DID.ID), nil)
-		err = credService.GetCredentials(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var getCredsResp router.GetCredentialsResponse
-		err = json.NewDecoder(w.Body).Decode(&getCredsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getCredsResp)
-
-		parsedCredsResponse, err := signing.ParseVerifiableCredentialFromJWT(getCredsResp.CredentialJWTs[0])
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredsResponse)
-
-		assert.Equal(tt, parsedCredResponse.ID, parsedCredsResponse.ID)
-		assert.Equal(tt, parsedCredResponse.CredentialSchema.ID, parsedCredsResponse.CredentialSchema.ID)
-
-		assert.Len(tt, getCredsResp.CredentialJWTs, 1)
-		assert.Equal(tt, parsedCredResponse.ID, parsedCredsResponse.ID)
-		assert.Equal(tt, parsedCredResponse.Issuer, parsedCredsResponse.Issuer)
-	})
-
-	t.Run("Test Get Credential By Subject", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		w := httptest.NewRecorder()
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		schemaID := "https://test-schema.com/name"
-		subjectID := "did:abc:456"
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: subjectID,
-			Schema:  schemaID,
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, resp.CredentialJWT)
-
-		parsedCredResponse, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredResponse)
-
-		w.Flush()
-
-		// get credential by subject id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credential?subject=%s", subjectID), nil)
-		err = credService.GetCredentials(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var getCredsResp router.GetCredentialsResponse
-		err = json.NewDecoder(w.Body).Decode(&getCredsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getCredsResp)
-
-		parsedCredsResponse, err := signing.ParseVerifiableCredentialFromJWT(getCredsResp.CredentialJWTs[0])
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredsResponse)
-
-		assert.Len(tt, getCredsResp.CredentialJWTs, 1)
-		assert.Equal(tt, parsedCredResponse.ID, parsedCredsResponse.ID)
-		assert.Equal(tt, parsedCredResponse.CredentialSubject[credsdk.VerifiableCredentialIDProperty], parsedCredsResponse.CredentialSubject[credsdk.VerifiableCredentialIDProperty])
-	})
-
-	t.Run("Test Delete Credential", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credService := testCredentialRouter(tt, bolt, keyStoreService)
-		didService := testDIDService(tt, bolt, keyStoreService)
-
-		issuerDID, err := didService.CreateDIDByMethod(did.CreateDIDRequest{
-			Method:  did.KeyMethod,
-			KeyType: crypto.Ed25519,
-		})
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, issuerDID)
-
-		createCredRequest := router.CreateCredentialRequest{
-			Issuer:  issuerDID.DID.ID,
-			Subject: "did:abc:456",
-			Data: map[string]interface{}{
-				"firstName": "Jack",
-				"lastName":  "Dorsey",
-			},
-			Expiry: time.Now().Add(24 * time.Hour).Format(time.RFC3339),
-		}
-		requestValue := newRequestValue(tt, createCredRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/credentials", requestValue)
-		w := httptest.NewRecorder()
-		err = credService.CreateCredential(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		parsedCredResponse, err := signing.ParseVerifiableCredentialFromJWT(*resp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedCredResponse)
-		credID := parsedCredResponse.ID
-
-		w.Flush()
-
-		// get credential by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credentials/%s", credID), nil)
-		err = credService.GetCredential(newRequestContextWithParams(map[string]string{"id": credID}), w, req)
-		assert.NoError(tt, err)
-
-		var getCredResp router.GetCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&getCredResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getCredResp)
-
-		parsedGetCred, err := signing.ParseVerifiableCredentialFromJWT(*getCredResp.CredentialJWT)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, parsedGetCred)
-
-		assert.Equal(tt, parsedCredResponse.ID, parsedGetCred.ID)
-
-		w.Flush()
-
-		// delete it
-		req = httptest.NewRequest(http.MethodDelete, fmt.Sprintf("https://ssi-service.com/v1/credentials/%s", credID), nil)
-		err = credService.DeleteCredential(newRequestContextWithParams(map[string]string{"id": credID}), w, req)
-		assert.NoError(tt, err)
-
-		w.Flush()
-
-		// get it back
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/credentials/%s", credID), nil)
-		err = credService.GetCredential(newRequestContextWithParams(map[string]string{"id": credID}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), fmt.Sprintf("could not get credential with id: %s", credID))
-	})
-}
-
-func TestManifestAPI(t *testing.T) {
-	t.Run("Test Create Manifest", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		// missing required field: Manifest
-		badManifestRequest := router.CreateManifestRequest{}
-
-		badRequestValue := newRequestValue(tt, badManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", badRequestValue)
-		w := httptest.NewRecorder()
-
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "invalid create manifest request")
-
-		// reset the http recorder
-		w.Flush()
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, resp.Manifest)
-		assert.Equal(tt, resp.Manifest.Issuer.ID, "did:abc:123")
-	})
-
-	t.Run("Test Get Manifest By ID", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		w := httptest.NewRecorder()
-
-		// get a manifest that doesn't exit
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/bad", nil)
-		err = manifestRouter.GetManifest(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "cannot get manifest without ID parameter")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// get a manifest with an invalid id parameter
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/bad", nil)
-		err = manifestRouter.GetManifest(newRequestContextWithParams(map[string]string{"id": "bad"}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not get manifest with id: bad")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		// get manifest by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/%s", resp.Manifest.ID), nil)
-		err = manifestRouter.GetManifest(newRequestContextWithParams(map[string]string{"id": resp.Manifest.ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getManifestResp router.GetManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&getManifestResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getManifestResp)
-		assert.Equal(tt, resp.Manifest.ID, getManifestResp.ID)
-	})
-
-	t.Run("Test Get Manifests", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		w := httptest.NewRecorder()
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		// get manifest by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests"), nil)
-		err = manifestRouter.GetManifests(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var getManifestsResp router.GetManifestsResponse
-		err = json.NewDecoder(w.Body).Decode(&getManifestsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getManifestsResp)
-		assert.Len(tt, getManifestsResp.Manifests, 1)
-		assert.Equal(tt, resp.Manifest.ID, getManifestsResp.Manifests[0].ID)
-	})
-
-	t.Run("Test Delete Manifest", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		w := httptest.NewRecorder()
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		w.Flush()
-
-		// get credential by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/%s", resp.Manifest.ID), nil)
-		err = manifestRouter.GetManifest(newRequestContextWithParams(map[string]string{"id": resp.Manifest.ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getManifestResp router.GetCredentialResponse
-		err = json.NewDecoder(w.Body).Decode(&getManifestResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getManifestResp)
-		assert.Equal(tt, resp.Manifest.ID, getManifestResp.ID)
-
-		w.Flush()
-
-		// delete it
-		req = httptest.NewRequest(http.MethodDelete, fmt.Sprintf("https://ssi-service.com/v1/manifests/%s", resp.Manifest.ID), nil)
-		err = manifestRouter.DeleteManifest(newRequestContextWithParams(map[string]string{"id": resp.Manifest.ID}), w, req)
-		assert.NoError(tt, err)
-
-		w.Flush()
-
-		// get it back
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/%s", resp.Manifest.ID), nil)
-		err = manifestRouter.GetManifest(newRequestContextWithParams(map[string]string{"id": resp.Manifest.ID}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), fmt.Sprintf("could not get manifest with id: %s", resp.Manifest.ID))
-	})
-
-	t.Run("Test Submit Application", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		// missing required field: Application
-		badManifestRequest := router.SubmitApplicationRequest{
-			RequesterDID: "id123",
-		}
-
-		badRequestValue := newRequestValue(tt, badManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests/applications", badRequestValue)
-		w := httptest.NewRecorder()
-
-		err = manifestRouter.SubmitApplication(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "invalid submit application request")
-
-		// reset the http recorder
-		w.Flush()
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, resp.Manifest)
-		assert.Equal(tt, resp.Manifest.Issuer.ID, "did:abc:123")
-
-		// good application request
-		createApplicationRequest := getValidApplicationRequest(resp.Manifest.ID, resp.Manifest.PresentationDefinition.InputDescriptors[0].ID)
-
-		applicationRequestValue := newRequestValue(tt, createApplicationRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests/applications", applicationRequestValue)
-		err = manifestRouter.SubmitApplication(newRequestContext(), w, req)
-
-		var appResp router.SubmitApplicationResponse
-		err = json.NewDecoder(w.Body).Decode(&appResp)
-		assert.NoError(tt, err)
-
-		assert.NotEmpty(tt, appResp.Response)
-		assert.Equal(tt, appResp.Response.ManifestID, resp.Manifest.ID)
-
-	})
-
-	t.Run("Test Get Application By ID and Get Applications", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-		w := httptest.NewRecorder()
-
-		// get a application that doesn't exit
-		req := httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/applications/bad", nil)
-		err = manifestRouter.GetApplication(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "cannot get application without ID parameter")
-
-		// reset recorder between calls
-		w.Flush()
-
-		// good manifest request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		// good application request
-		createApplicationRequest := getValidApplicationRequest(resp.Manifest.ID, resp.Manifest.PresentationDefinition.InputDescriptors[0].ID)
-
-		applicationRequestValue := newRequestValue(tt, createApplicationRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests/applications", applicationRequestValue)
-		err = manifestRouter.SubmitApplication(newRequestContext(), w, req)
-
-		var appResp router.SubmitApplicationResponse
-		err = json.NewDecoder(w.Body).Decode(&appResp)
-		assert.NoError(tt, err)
-
-		// get response by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/responses/%s", appResp.Response.ID), nil)
-		err = manifestRouter.GetResponse(newRequestContextWithParams(map[string]string{"id": appResp.Response.ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getResponseResponse router.GetResponseResponse
-		err = json.NewDecoder(w.Body).Decode(&getResponseResponse)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getResponseResponse)
-		assert.Equal(tt, appResp.Response.ID, getResponseResponse.Response.ID)
-
-		// get all responses
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/responses", nil)
-		err = manifestRouter.GetResponses(newRequestContext(), w, req)
-
-		var getResponsesResp router.GetResponsesResponse
-		err = json.NewDecoder(w.Body).Decode(&getResponsesResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getResponsesResp)
-
-		assert.Len(tt, getResponsesResp.Responses, 1)
-
-		// get all applications
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/applications", applicationRequestValue)
-		err = manifestRouter.GetApplications(newRequestContext(), w, req)
-
-		var getApplicationsResp router.GetApplicationsResponse
-		err = json.NewDecoder(w.Body).Decode(&getApplicationsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getApplicationsResp)
-
-		assert.Len(tt, getApplicationsResp.Applications, 1)
-
-		// get application by id
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/applications/%s", getApplicationsResp.Applications[0].ID), nil)
-		err = manifestRouter.GetApplication(newRequestContextWithParams(map[string]string{"id": getApplicationsResp.Applications[0].ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getApplicationResponse router.GetApplicationResponse
-		err = json.NewDecoder(w.Body).Decode(&getApplicationResponse)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getApplicationResponse)
-		assert.Equal(tt, getApplicationsResp.Applications[0].ID, getApplicationResponse.ID)
-	})
-
-	t.Run("Test Delete Application", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, _ := testManifest(tt, bolt, keyStoreService, credentialService)
-
-		// good manifest request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		w := httptest.NewRecorder()
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		// good application request
-		createApplicationRequest := getValidApplicationRequest(resp.Manifest.ID, resp.Manifest.PresentationDefinition.InputDescriptors[0].ID)
-
-		applicationRequestValue := newRequestValue(tt, createApplicationRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests/applications", applicationRequestValue)
-		err = manifestRouter.SubmitApplication(newRequestContext(), w, req)
-
-		var appResp router.SubmitApplicationResponse
-		err = json.NewDecoder(w.Body).Decode(&appResp)
-		assert.NoError(tt, err)
-
-		// get all applications
-		req = httptest.NewRequest(http.MethodGet, "https://ssi-service.com/v1/manifests/applications", applicationRequestValue)
-		err = manifestRouter.GetApplications(newRequestContext(), w, req)
-
-		var getApplicationsResp router.GetApplicationsResponse
-		err = json.NewDecoder(w.Body).Decode(&getApplicationsResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getApplicationsResp)
-
-		assert.Len(tt, getApplicationsResp.Applications, 1)
-
-		// get the application
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/applications/%s", getApplicationsResp.Applications[0].ID), nil)
-		err = manifestRouter.GetApplication(newRequestContextWithParams(map[string]string{"id": getApplicationsResp.Applications[0].ID}), w, req)
-		assert.NoError(tt, err)
-
-		var getApplicationResp router.GetApplicationResponse
-		err = json.NewDecoder(w.Body).Decode(&getApplicationResp)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, getApplicationResp)
-		assert.Equal(tt, resp.Manifest.ID, getApplicationResp.Application.ManifestID)
-
-		// delete the application
-		req = httptest.NewRequest(http.MethodDelete, fmt.Sprintf("https://ssi-service.com/v1/manifests/applications/%s", getApplicationResp.Application.ID), nil)
-		err = manifestRouter.DeleteApplication(newRequestContextWithParams(map[string]string{"id": getApplicationResp.Application.ID}), w, req)
-		assert.NoError(tt, err)
-
-		w.Flush()
-
-		// get it back
-		req = httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/manifests/applications/%s", appResp.Response.ID), nil)
-		err = manifestRouter.GetApplication(newRequestContextWithParams(map[string]string{"id": appResp.Response.ID}), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), fmt.Sprintf("could not get application with id: %s", appResp.Response.ID))
-	})
-}
-
-func TestDWNAPI(t *testing.T) {
-	t.Run("Test DWN Publish Manifest", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService := testKeyStoreService(tt, bolt)
-		credentialService := testCredentialService(tt, bolt, keyStoreService)
-		manifestRouter, manifestService := testManifest(tt, bolt, keyStoreService, credentialService)
-		dwnService := testDWNRouter(tt, bolt, keyStoreService, manifestService)
-
-		w := httptest.NewRecorder()
-
-		// good request
-		createManifestRequest := getValidManifestRequest()
-
-		requestValue := newRequestValue(tt, createManifestRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/manifests", requestValue)
-		err = manifestRouter.CreateManifest(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		var resp router.CreateManifestResponse
-		err = json.NewDecoder(w.Body).Decode(&resp)
-		assert.NoError(tt, err)
-
-		w = httptest.NewRecorder()
-
-		dwnRequest := router.PublishManifestRequest{ManifestID: "WA-DL-CLASS-A"}
-		dwnRequestValue := newRequestValue(tt, dwnRequest)
-		dwnReq := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/dwn/manifests", dwnRequestValue)
-		err = dwnService.PublishManifest(newRequestContext(), w, dwnReq)
-
-		assert.Error(tt, err)
-		assert.ErrorContains(tt, err, "unsupported protocol scheme")
-	})
-}
-
-func TestKeyStoreAPI(t *testing.T) {
-	t.Run("Test Store Key", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreRouter, _ := testKeyStore(tt, bolt)
-		w := httptest.NewRecorder()
-
-		// bad key type
-		badKeyStoreRequest := router.StoreKeyRequest{
-			ID:               "test-kid",
-			Type:             "bad",
-			Controller:       "me",
-			Base58PrivateKey: "bad",
-		}
-		badRequestValue := newRequestValue(tt, badKeyStoreRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/keys", badRequestValue)
-		err = keyStoreRouter.StoreKey(newRequestContext(), w, req)
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not store key: test-kid, unsupported key type: bad")
-
-		// reset the http recorder
-		w.Flush()
-
-		// store a valid key
-		_, privKey, err := crypto.GenerateKeyByKeyType(crypto.Ed25519)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, privKey)
-
-		privKeyBytes, err := crypto.PrivKeyToBytes(privKey)
-		assert.NoError(tt, err)
-
-		// good request
-		storeKeyRequest := router.StoreKeyRequest{
-			ID:               "did:test:me#key-1",
-			Type:             crypto.Ed25519,
-			Controller:       "did:test:me",
-			Base58PrivateKey: base58.Encode(privKeyBytes),
-		}
-		requestValue := newRequestValue(tt, storeKeyRequest)
-		req = httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/keys", requestValue)
-		err = keyStoreRouter.StoreKey(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-	})
-
-	t.Run("Test Get Key Details", func(tt *testing.T) {
-		bolt, err := storage.NewBoltDB()
-
-		// remove the db file after the test
-		tt.Cleanup(func() {
-			_ = bolt.Close()
-			_ = os.Remove(storage.DBFile)
-		})
-
-		keyStoreService, _ := testKeyStore(tt, bolt)
-		w := httptest.NewRecorder()
-
-		// store a valid key
-		_, privKey, err := crypto.GenerateKeyByKeyType(crypto.Ed25519)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, privKey)
-
-		privKeyBytes, err := crypto.PrivKeyToBytes(privKey)
-		assert.NoError(tt, err)
-
-		// good request
-		keyID := "did:test:me#key-2"
-		controller := "did:test:me"
-		storeKeyRequest := router.StoreKeyRequest{
-			ID:               keyID,
-			Type:             crypto.Ed25519,
-			Controller:       controller,
-			Base58PrivateKey: base58.Encode(privKeyBytes),
-		}
-		requestValue := newRequestValue(tt, storeKeyRequest)
-		req := httptest.NewRequest(http.MethodPut, "https://ssi-service.com/v1/keys", requestValue)
-		err = keyStoreService.StoreKey(newRequestContext(), w, req)
-		assert.NoError(tt, err)
-
-		// get it back
-		getRecorder := httptest.NewRecorder()
-		getReq := httptest.NewRequest(http.MethodGet, fmt.Sprintf("https://ssi-service.com/v1/keys/%s", keyID), nil)
-		err = keyStoreService.GetKeyDetails(newRequestContextWithParams(map[string]string{"id": keyID}), getRecorder, getReq)
-		assert.NoError(tt, err)
-
-		var resp router.GetKeyDetailsResponse
-		err = json.NewDecoder(getRecorder.Body).Decode(&resp)
-		assert.NoError(tt, err)
-		assert.Equal(tt, keyID, resp.ID)
-		assert.Equal(tt, controller, resp.Controller)
-		assert.Equal(tt, crypto.Ed25519, resp.Type)
-	})
 }
 
 func newRequestValue(t *testing.T, data interface{}) io.Reader {
